requirements.txt
GitPython==3.1.29
Jinja2==3.1.2
PyYAML==6.0
celery[redis]==5.2.7
cliff==4.1.0
deepdiff==6.2.2
fastapi==0.88.0
flower==1.2.0
hiredis==2.1.0
kombu==5.2.4
loguru==0.6.0
napalm==4.0.0
netmiko==4.1.2
openstacksdk==0.103.0
pottery==3.0.0
pydantic==1.10.2
pynetbox==7.0.0
sushy==4.4.1
tabulate==0.9.0
transitions==0.9.0
uvicorn[standard]==0.20.0
watchdog==2.2.0
